+++ code/render/batching.cpp
@@ -28,7 +28,7 @@ void primitive_batch::add_point_sprite(batch_vertex *p)
 	Vertices.push_back(*p);
 }
 
-int primitive_batch::load_buffer(batch_vertex* buffer, int n_verts)
+size_t primitive_batch::load_buffer(batch_vertex* buffer, size_t n_verts)
 {
 	size_t verts_to_render = Vertices.size();
